--- a/rpcs3/Input/keyboard_pad_handler.cpp
+++ b/rpcs3/Input/keyboard_pad_handler.cpp
@@ -750,7 +750,7 @@ std::string keyboard_pad_handler::native_scan_code_to_string(int native_scan_cod
 
 bool keyboard_pad_handler::bindPadToDevice(std::shared_ptr<Pad> pad, const std::string& device, u8 player_id)
 {
-	if (device != pad::keyboard_device_name)
+	if (!pad || device != pad::keyboard_device_name)
 		return false;
 
 	m_pad_configs[player_id].from_string(g_cfg_input.player[player_id]->config.to_string());