@sampletext32 2022-01-29
Hey, @Megamouse, I have 2 DualSense controllers, so how can I help you with testing this one?