@Megamouse 2022-01-29
You could check if you can feel some sort of input lag every 2 seconds on master and if it's fixed with this PR.
Keep in mind that it's probably gonna be hard to notice