@Megamouse 2022-01-31
I'll only keep the hid enumeration multithreaded for now.
I'll add multithreading for pad handlers in a follow-up PR.